refactor: CommentService update delete 메소드 수정 수정은 findById, 삭제는 findByIdWithPost로 변경

+++ src/main/java/com/boilerplate/boilerplate/domain/post/service/PostService.java
@@ -54,7 +54,11 @@ public class PostService {
 
     @PreAuthorize("hasRole('ADMIN') or @postSecurityChecker.isPostOwner(#postId)")
     public void delete(Long postId) {
+        if (!postRepository.existsById(postId)) {
+            throw new EntityNotFoundException(PostError.POST_NOT_EXIST.getMessage());
+        }
         postRepository.deleteById(postId);
+
     }
 
     @Transactional(readOnly = true)
